barebox: add package name argument
No functional changes: Adds an argument to the barebox-package function to make
the uppercase package name configurable. This is needed to support a 2nd config
build. This results in the following renaming: 'BAREBOX' -> '$(1)'

KConfig package source variables are not yet extracted under the same argument:
The barebox and barebox-2 packages will have common package source definitions.