barebox: add package origin namespace argument
No functional changes: Adds an argument to the barebox-package function to make
the package origin namespace configurable. The package origin namespace are all
KConfig variables that determine which sources will be built. This includes the
package version, URL, patches, etc. This is needed to support a 2nd
config build. The result is the following renaming: 'BAREBOX' -> '$(2)'