barebox: extract package name argument
No functional changes: Extracts an argument to the inner-barebox-package
function to automatically determine the uppercase package name. This is needed
to support a 2nd config build. This results in the following renaming:
  'BAREBOX' -> '$(1)'

All barebox packages are meant to be built from the same sources, so related
KConfig variables (origin, version and patch directory) are not extracted.